add:role menu relation events

+++ LogisticsManagementSystem.Application/Roles/Commands/DeleteRole/DeleteRoleCommandHandler.cs
@@ -4,7 +4,7 @@ using MediatR;
 
 namespace LogisticsManagementSystem.Application;
 
-public class DeleteRoleCommandHandler(IRoleRepository _roleRepository) : IRequestHandler<DeleteRoleCommand, ErrorOr<Deleted>>
+public class DeleteRoleCommandHandler(IRoleRepository _roleRepository, IUserRolesRepository _userRolesRepository) : IRequestHandler<DeleteRoleCommand, ErrorOr<Deleted>>
 {
     public async Task<ErrorOr<Deleted>> Handle(DeleteRoleCommand command, CancellationToken cancellationToken)
     {
@@ -12,7 +12,16 @@ public class DeleteRoleCommandHandler(IRoleRepository _roleRepository) : IReques
         if (role is null)
             return Error.NotFound(description: "用户不存在");
 
+        var roles = await _userRolesRepository.GetListByRoleIdAsync(role.Id, cancellationToken);
+
+        if (roles.Any())
+        {
+            return Error.Conflict(description: "有用户设置当前的角色");
+        }
+        role.DeleteRoleMenus(role.Id);
+        role.SetDeletedAt();
         await _roleRepository.DeleteAsync(role, cancellationToken);
+
         return Result.Deleted;
     }
 }
